requirements.txt
flask==2.3.2
flask-cors==3.0.10
tensorflow==2.10.0
opencv-python==4.7.0.72
numpy==1.23.5
pillow==9.5.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
gunicorn==20.1.0